// src/app/youtube/components/search/models/search-item.model.ts
export interface ISearchItem {
  kind: string;
  etag: string;
  id: {
    kind: string;
    videoId: string;
  };
  snippet: {
    publishedAt: string;
    channelId: string;
    title: string;
    description: string;
    thumbnails: {
      default: VideoThumbnails;
      medium: VideoThumbnails;
      high: VideoThumbnails;
      standard: VideoThumbnails;
      maxres: VideoThumbnails;
    };
    channelTitle: string;
    tags: string[];
    categoryId: string;
    liveBroadcastContent: string;
    localized: {
      title: string;
      description: string;
    };
    defaultAudioLanguage: string;
  };
  statistics: {
    viewCount: string;
    likeCount: string;
    dislikeCount: string;
    favoriteCount: string;
    commentCount: string;
  };
}

type VideoThumbnails = {
  url: string;
  width: number;
  height: number;
};